Add sorting concern to controllers
